+++ packages/0x.js/CHANGELOG.md
@@ -4,6 +4,7 @@
 
     * Validate and lowercase all addresses in public methods (#373)
     * Improve validation to force passing contract addresses on private networks (#385)
+    * Change `LogErrorContractEventArgs.errorId` type from `BigNumber` to `number` (#413)
     * Rename all public `_unsubscribeAll` methods to `unsubscribeAll` (#415)
 
 ## v0.32.2 - _February 9, 2018_